Move artifact transform test to `*DependencyResolutionIntegrationTest`
