+++ subprojects/instant-execution/src/integTest/groovy/org/gradle/instantexecution/InstantExecutionDependencyResolutionIntegrationTest.groovy
@@ -17,6 +17,7 @@
 package org.gradle.instantexecution
 
 import org.gradle.integtests.resolve.transform.ArtifactTransformTestFixture
+import spock.lang.Ignore
 import spock.lang.Issue
 
 class InstantExecutionDependencyResolutionIntegrationTest extends AbstractInstantExecutionIntegrationTest implements ArtifactTransformTestFixture {
@@ -148,7 +149,7 @@ class InstantExecutionDependencyResolutionIntegrationTest extends AbstractInstan
         result.assertTaskOrder(":b:producer", ":resolve")
         result.assertTaskNotSkipped(":a:producer")
         result.assertTaskSkipped(":b:producer")
-        result.assertTaskNotSkipped(    ":resolve")
+        result.assertTaskNotSkipped(":resolve")
         file('out.txt').text == "12,10"
     }
 
@@ -563,5 +564,106 @@ class InstantExecutionDependencyResolutionIntegrationTest extends AbstractInstan
         outputContains("processing b.jar.red using []")
         outputContains("processing a.jar.red using [b.jar.red, c.jar.red]")
         outputContains("result = [a.jar.red.green, b.jar.red.green, c.jar.red.green]")
+    }
+
+    @Ignore("wip")
+    def 'transform action is re-executed when input artifact changes'() {
+        given:
+        buildKotlinFile '''
+
+abstract class Summarize : TransformAction<TransformParameters.None> {
+
+    @get:InputArtifact
+    abstract val inputArtifact: Provider<FileSystemLocation>
+
+    override fun transform(outputs: TransformOutputs) {
+        val inputFile = inputArtifact.get().asFile
+        println("Transforming ${inputFile.name}...")
+        outputs.file("${inputFile.nameWithoutExtension}-summary.txt").run {
+            writeText("${inputFile.name}: ${inputFile.length()}")
+        }
+    }
+}
+
+val summarized = Attribute.of("summarized", Boolean::class.javaObjectType)
+dependencies {
+    attributesSchema {
+        attribute(summarized)
+    }
+    artifactTypes.create("txt") {
+        attributes.attribute(summarized, false)
+    }
+    registerTransform(Summarize::class) {
+        from.attribute(summarized, false)
+        to.attribute(summarized, true)
+    }
+}
+
+val sourceFiles by configurations.creating {
+    isCanBeConsumed = false
+    isCanBeResolved = false
+}
+
+val summarizedFiles by configurations.creating {
+    extendsFrom(sourceFiles)
+    isCanBeConsumed = false
+    isCanBeResolved = true
+    attributes {
+        attribute(summarized, true)
+    }
+}
+
+abstract class CombineSummaries : DefaultTask() {
+
+    @get:InputFiles
+    abstract val inputSummaries: ConfigurableFileCollection
+
+    @get:OutputFile
+    abstract val outputFile: RegularFileProperty
+
+    @TaskAction
+    fun summarize() {
+        outputFile.get().asFile.run {
+            parentFile.mkdirs()
+            writeText(summaryString())
+        }
+    }
+
+    private
+    fun summaryString() = inputSummaries.files.joinToString(separator = "\\n") { it.readText() }
+}
+
+tasks.register<CombineSummaries>("summarize") {
+    inputSummaries.from(summarizedFiles)
+    outputFile.set(layout.buildDirectory.file("summary.txt"))
+}
+
+dependencies {
+    sourceFiles(files("input.txt"))
+}
+'''
+        def inputFile = file('input.txt').tap { write("the input file") }
+        def outputFile = file('build/summary.txt')
+        def expectedOutput = "input.txt: ${inputFile.length()}"
+        def instant = newInstantExecutionFixture()
+
+        when:
+        instantRun 'summarize'
+
+        then:
+        instant.assertStateStored()
+        outputFile.text == expectedOutput
+        outputContains 'Transforming input.txt...'
+        result.assertTaskExecuted ':summarize'
+
+        when: 'input file changes'
+        inputFile.text = inputFile.text.reverse()
+        instantRun 'summarize'
+
+        then:
+        instant.assertStateLoaded()
+        outputFile.text == expectedOutput
+        outputContains 'Transforming input.txt...'
+        result.assertTaskExecuted ':summarize'
     }
 }
